# [SofaDefaultType] Speedup MapMapSparseMatrix