@fredroy 2022-01-28
> It means the data is no longer processed in a predictable order. In most cases, it does not change anything since the data structure is converted to another. 

I dont see any case where order is needed as you said, but it could be that someone somewhere would need it?
Should not we add a template parameter on the map type (and set it to default unordered_map) ?
I am wondering also if we should make this PR breaking by itself (breaking in the meaning of different behavior.

Otherwise when I see the 1:10 ratio, I can see this is needed 😏