@alxbilger 2022-01-28
@fredroy I had the same idea to add a template parameter. However, this class is used as a main type in all DataTypes (Vec3d, Rigid3d etc). Which means there is no mechanism to choose which version of the map (ordered or unordered) will be used. The type is imposed to all the component using the same DataTypes.
Also, this class days are counted. It has to be replaced by an efficient sparse matrix data structure. But when?
I am ready to add the template parameter if you think it is necessary.