@fredroy 2022-01-28
> I am ready to add the template parameter if you think it is necessary.

Personally I dont really care (I like avoiding template if possible 😏) and I dont know if this structure is really used outside SOFA...