@EulalieCoevoet 2022-02-11
Hello @fredroy, @hugtalbot, @alxbilger 

This PR seems to mess up with the mapping of constraints sometimes. It seems to fail when the constraint is under two mappings? Does that make sense? I can make a minimal example but with BeamAdapter and SoftRobots... Tell me if you're interested anyway. 

It could be usefull to add the breaking label.