@alxbilger 2022-02-16
I was surprised that it did not break anything. Since it does not bring an extraordinary speed up, I suggest to revert this PR for the time being. 
One day, this matrix structure will have to be replaced